version 4.3.0

+++ inst/include/bigmemory/bigmemoryDefines.h
@@ -15,7 +15,8 @@
 #ifdef __intptr_t_defined
   typedef intptr_t index_type;
 #else
-  typedef long long index_type;
+  // We don't have 64-bit indexing.  Build with 32.
+  typedef long index_type;
 #endif
 
 extern "C"
